refactory. 이름, 시도 횟수 검증 클래스 분할
